requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitsandbytes==0.46.1
botorch==0.14.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
distro==1.9.0
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
gitignore_parser==0.1.12
GitPython==3.1.45
gpytorch==1.14
graphviz==0.21
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
jaxtyping==0.3.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
linear-operator==0.6
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mixture-of-experts==0.2.3
mpmath==1.3.0
multidict==6.6.3
multipledispatch==1.0.0
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.5.1
openai==1.98.0
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.34.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
token-count==0.2.1
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wadler_lindig==0.1.7
wandb==0.21.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
